Fix starter upper bound (integer part was three times greater for some items), add space upper bound
The space upper bound currently is restricted to integers, so the actual value is somewhere between it and 1 less than it (which the next commit should solve).
Compared to before (where it was returning naive floats that assumed things like that exactly 1/4th of a seller is used per item/second, and 2/3rds of a machine (1/3rd of a starter-and-triple-splitter configuration) per processed material, so was only asymptotically correct to a very naive upper bound), it accounts for the number of starters consumed properly now, and assumes the number of squares reserved for splitters is at least the number by which the raw material's machines into which they're distributed exceed the number of starters producing it, ceilingdiv 2 (a regular splitter is still necessary when a triple one isn't), summed over raw material types. It doesn't account for the use of splitter synchronisation (ie. when two starters both need to input 1 raw material/second into each of the same two crafters, you could solve it with a single 2-way splitter set to [1,1] instead of two), and it doesn't account for the necessary rollers in distributing them, because to do so would require either iterating over all possibilities (which is infeasible) or making assumptions (however, from each of its proofs, you could arrange machines as nodes in a graph, each connected to at most four neighbours but not guaranteedly (and almost certainly not) immersible in a grid without intermediate squares (rollers)).
Even so, as well as all processed materials, it has found lower upper bounds on production rates for the circuit, engine, antenna, battery, headphones, drill, power supply, speakers, radio, microwave, generator and electric board than the starter limit alone can provide.
I think, at least in this arbitrary non-immersible graph scheme, there is no instance wherein removal of splitters is .
If anyone knows of a design exceeding either upper limit this provides, I will be disheartened but interested to know, but I haven't found any yet.
Perhaps generating these graphs and simulating them (like my tablebase program does) could allow humans to have ideas for splitter-synchronisationless designs more easily, where all innovations like uses of selectors to cross edges are applied in the immersion process (though, being an Assembly Line patzer (as it were), I don't know whether the people who make very complicated designs perform an equivalent process in their heads already).